Moved PlanningTable into PlanningOverview as it is only used there.
