+++ src/views/PlanningOverview/PlanningGridColumn.tsx
@@ -1,5 +1,5 @@
 
-import { type Planning as PlanningType } from '@/components/PlanningTable/data/schema'
+import { type Planning as PlanningType } from '@/views/PlanningOverview/PlanningTable/data/schema'
 import { useRegistry } from '@/hooks'
 import { SectorBadge } from '@/components/DataItem/SectorBadge'
 import { StatusIndicator } from '@/components/DataItem/StatusIndicator'
